font awesome crud icons in mn-grid

+++ app/assets/javascripts/mohican/directives/mnf-delete-grid/directive.js
@@ -14,9 +14,11 @@ angular.module('mohican.directives')
 
         link: function(scope, elem, attr, mnfFormGridCtrl) {
           scope.confirmDelete = function() {
-            if($window.confirm('Are you sure that you want to permanently delete document?')) {
-              scope.mnfDoc.delete();
-              mnfFormGridCtrl.currentMnfDoc = null;
+            if(scope.mnfDoc._state === 'ready') {
+              if($window.confirm('Are you sure that you want to permanently delete document?')) {
+                scope.mnfDoc.delete();
+                mnfFormGridCtrl.currentMnfDoc = null;
+              }
             }
           };
         }
